Cleaned up Saturday on a Sundee
Added all the logic from Saturday to the GenerateAndSetKitchen method.
The solution is becoming very specific to Maseto's, and with more rules
I think I can make it more typical to the regular distribution. Because
of the order of selection. Sandwichmakers have a higher chance to get
shifts, but this could be corrected.