Starting to Implement busy_persons()
It's causing a lot of ValueErrors which means it's running out of people
in its random.sample() calls. It might fix it to separate out all the
different generate position calls and order them so its most likely to
succeed. That is, from the smallest set (Sandwiches) to the largest
(Helper)